% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdge-shiny-modules.R
\name{fdgeViewServer}
\alias{fdgeViewServer}
\title{Shiny module that interacts with result of fdgeRun module}
\usage{
fdgeViewServer(
  id,
  rfds,
  dgeres,
  ...,
  feature_selection = session$ns("volcano"),
  sample_selection = session$ns("samples"),
  debug = FALSE
)
}
\arguments{
\item{dgeres}{The result from calling \code{\link[=fdge]{fdge()}}, or \code{\link[=fdgeRun]{fdgeRun()}}.}
}
\description{
Provides an interactive view over the result of running the \code{\link[=fdgeRun]{fdgeRun()}}
module.
}
